[upd] Move time calculation to thread, select more accurate timer
Changed the place where calculation of time available is. Do it in the separate thread.
Time available time calculation was longer than 100 mSec before so because the timer is every 100 mSec the actual timer was longer that time in reality.

moved some of the processing in the second thread and select higher accuracy timer.